Features added to V4 of the Project
The application has optional features that can be activated via switches in the interface. The "Keep the couple together" feature, when activated, modifies the distribution logic to ensure that players "Neto" and "Luisa," if both are among the 12 selected, are placed on the same team. The algorithm does this by treating them as a priority unit, placing them together on the lowest-scoring team before proceeding with the distribution of the remaining 10 players.

To promote an inclusive environment and avoid embarrassment, the application defaults to hiding all players' star ratings. This measure aims to protect the well-being of participants, especially those with lower ratings. A second optional feature, activated by a "Show Ratings" switch, allows the match organizer, if deemed appropriate, to make the stars visible in the interface, both on the selection screen and in the final team results. This inverted logic prioritizes player privacy and comfort.