Admin_Settings
This management page was developed as an administrative tool to facilitate the maintenance of the player database for the "Volleyball Times Draw" application. It allows group leaders or organizers to add new participants, update the technical ratings of existing players, or remove members who are no longer part of the group, all through an intuitive visual interface. The main objective is to generate an updated JSON data structure, which will be used to feed the main application, ensuring that times draws are always based on the most recent and accurate information.